base.html: bugfix for not always being passed JSON data
The registration template currently doesn't get passed the JSON
data. This is something that should eventually be fixed. The
favicon was reliant on the data and was causing an exception
when favicon wasn't supplied. Add an if/else into the base template
to either use the passed favicon or fall back to the default heart.

Also fix a similar issue with the explore header, as we can't
guarantee we'll be passed the explore title. Same goes
for the gride/broom initials, but when those are none they
just don't show up which is fine.

Also bump the bugfix version